Version 3.0
bootstrap.sh:
- Add function to control include of lib and files;
- Sourcing of dependencies file (pom.sh) moved from shpm.sh to
bootstrap.sh

shpm.sh:
- add operations to clean target, install and uninstall in local repo

+++ src/main/sh/shpm.sh
@@ -1,10 +1,10 @@
 #!/bin/bash
 
-source ../../../bootstrap.sh  
+source ../../../bootstrap.sh
 
-source $ROOT_DIR_PATH/pom.sh
-
-BOOTSTRAP_FILENAME="bootstrap.sh"
+shpm_log() {
+	echo "$1"
+}
 
 print_help() {
   
@@ -15,11 +15,14 @@ print_help() {
 	echo "  shpm [OPTION]"
 	echo ""
 	echo "OPTIONS:"
-    echo "  update                Download dependencies in $LIB_DIR_SUBPATH folder"
+    echo "  update                Download dependencies in local repository $LIB_DIR_SUBPATH"
+	echo "  clean                 Clean $TARGET_DIR_PATH folder"
     echo "  test                  Run tests in $TEST_DIR_SUBPATH folder"
-    echo "  build                 Create compressed file in $TARGET_DIR_PATH folder"        
+    echo "  build                 Create compressed file in $TARGET_DIR_PATH folder"
+	echo "  install               Install in local repository $LIB_DIR_SUBPATH"            
     echo "  publish               Publish compressed sh in repository"
     echo "  autoupdate            Update itself"
+	echo "  uninstall             Remove from local repository $LIB_DIR_SUBPATH"
 	echo ""
 	echo "EXAMPLES:"
 	echo "  ./shpm update"
@@ -34,9 +37,12 @@ run_sh_pm() {
 
 	local UPDATE=false
 	local TEST=false
+	local CLEAN=false
 	local BUILD=false
+	local INSTALL=false	
 	local PUBLISH=false
 	local AUTOUPDATE=false
+	local UNINSTALL=false
 	
 	if [ $# -eq 0 ];  then
 		print_help
@@ -52,9 +58,17 @@ run_sh_pm() {
 			if [[ "$ARG" == "test" ]];  then
 				TEST="true"
 			fi
+			
+			if [[ "$ARG" == "clean" ]];  then
+				CLEAN="true"
+			fi
 		
 			if [[ "$ARG" == "build" ]];  then
 				BUILD="true"
+			fi
+			
+			if [[ "$ARG" == "install" ]];  then
+				INSTALL="true"
 			fi
 			
 			if [[ "$ARG" == "publish" ]];  then
@@ -63,7 +77,11 @@ run_sh_pm() {
 			
 			if [[ "$ARG" == "autoupdate" ]];  then
 				AUTOUPDATE="true"
-			fi				
+			fi
+			
+			if [[ "$ARG" == "uninstall" ]];  then
+				UNINSTALL="true"
+			fi
 		done
 	fi
 	
@@ -72,12 +90,20 @@ run_sh_pm() {
 		update_dependencies	
 	fi
 	
-	if [[ "$BUILD" == "true" ]];  then
-		build_release
+	if [[ "$CLEAN" == "true" ]];  then
+		clean_release
 	fi
 	
 	if [[ "$TEST" == "true" ]];  then
 		run_all_tests
+	fi
+	
+	if [[ "$BUILD" == "true" ]];  then
+		build_release
+	fi
+	
+	if [[ "$INSTALL" == "true" ]];  then
+		install_release
 	fi
 	
 	if [[ "$PUBLISH" == "true" ]];  then	
@@ -87,28 +113,83 @@ run_sh_pm() {
 	if [[ "$AUTOUPDATE" == "true" ]];  then	
 		auto_update
 	fi
+	
+	if [[ "$UNINSTALL" == "true" ]];  then
+		uninstall_release
+	fi	
+}
+
+clean_release() {
+	if [[ ! -z "$TARGET_DIR_PATH" && -d "$TARGET_DIR_PATH" ]]; then
+		cd "$TARGET_DIR_PATH"
+		rm *.tar.gz 2> /dev/null
+	fi
 }
 
 update_dependencies() {
-
 	local HOST=${REPOSITORY[host]}
-	local PORT=${REPOSITORY[port]}	
-			
-	for DEP_ARTIFACT_ID in "${!DEPENDENCIES[@]}"; do 
+	local PORT=${REPOSITORY[port]}
 	
+	for DEP_ARTIFACT_ID in "${!DEPENDENCIES[@]}"; do 
 		update_dependency $DEP_ARTIFACT_ID
-		
 	done
 	
 	cd $ROOT_DIR_PATH
+}
 
+uninstall_release () {
+
+	clean_release
+	build_release
+	
+	local TARGET_FOLDER=$ARTIFACT_ID"-"$VERSION
+	local TGZ_FILE=$TARGET_FOLDER".tar.gz"
+	local TGZ_FILE_PATH=$TARGET_DIR_PATH/$TGZ_FILE
+	
+	local ACTUAL_DIR=$(pwd)
+	
+	cd $LIB_DIR_PATH/
+	rm *".tar.gz" 2> /dev/null	
+	
+	if [[ -d  $LIB_DIR_PATH/$TARGET_FOLDER ]]; then
+		# evict rm -rf!
+		mv $LIB_DIR_PATH/$TARGET_FOLDER /tmp 2> /dev/null
+		local TIMESTAMP=$( date +"%Y%m%d_%H%M%S_%N" )			
+		mv /tmp/$TARGET_FOLDER /tmp/$TARGET_FOLDER"_"$TIMESTAMP
+	fi	
+	
+	cd "$ACTUAL_DIR"
+}
+
+install_release () {
+
+	clean_release
+	build_release
+	uninstall_release
+	
+	local TARGET_FOLDER=$ARTIFACT_ID"-"$VERSION
+	local TGZ_FILE=$TARGET_FOLDER".tar.gz"
+	local TGZ_FILE_PATH=$TARGET_DIR_PATH/$TGZ_FILE
+	
+	local ACTUAL_DIR=$(pwd)
+	
+	cd $LIB_DIR_PATH/
+	
+	cp $TGZ_FILE_PATH $LIB_DIR_PATH/	
+	
+	tar -xzf $TGZ_FILE
+		
+	rm -f $TGZ_FILE
+	
+	cd "$ACTUAL_DIR"
 }
 
 update_dependency() {
 
         local DEP_ARTIFACT_ID=$1
-
-		DEP_VERSION="${DEPENDENCIES[$DEP_ARTIFACT_ID]}"		
+		local DEP_VERSION="${DEPENDENCIES[$DEP_ARTIFACT_ID]}"
+				
+		shpm_log "Update $DEP_ARTIFACT_ID to $DEP_VERSION: Start"
 		
 		if [[ ! -d $LIB_DIR_PATH ]]; then
 		  mkdir -p $LIB_DIR_PATH
@@ -116,11 +197,12 @@ update_dependency() {
 		
 		DEP_FILENAME=$DEP_ARTIFACT_ID"-"$DEP_VERSION".tar.gz"
 		
-		echo "http://$HOST:$PORT/sh-archiva/get/snapshot/$GROUP_ID/$DEP_ARTIFACT_ID/$DEP_VERSION > $LIB_DIR_PATH/$DEP_FILENAME"
-		curl  http://$HOST:$PORT/sh-archiva/get/snapshot/$GROUP_ID/$DEP_ARTIFACT_ID/$DEP_VERSION > $LIB_DIR_PATH/$DEP_FILENAME
+		shpm_log " - Downloading $DEP_FILENAME ..."
+		curl -s  http://$HOST:$PORT/sh-archiva/get/snapshot/$GROUP_ID/$DEP_ARTIFACT_ID/$DEP_VERSION > $LIB_DIR_PATH/$DEP_FILENAME 
 		
 		cd $LIB_DIR_PATH/
 		
+		shpm_log " - Extracting $DEP_FILENAME into $LIB_DIR_PATH/$DEP_FILENAME ..."
 		if [[ -d  $LIB_DIR_PATH/$DEP_ARTIFACT_ID ]]; then
 			# evict rm -rf!
 			mv $LIB_DIR_PATH/$DEP_ARTIFACT_ID /tmp 2> /dev/null
@@ -128,27 +210,36 @@ update_dependency() {
 			mv /tmp/$DEP_ARTIFACT_ID /tmp/$DEP_ARTIFACT_ID"_"$TIMESTAMP
 		fi
 		
-		tar -xvzf $DEP_FILENAME
+		tar -xzf $DEP_FILENAME
 		
 		rm -f $DEP_FILENAME
 		
 		# if update a sh-pm
 		if [[ "$DEP_ARTIFACT_ID" == "sh-pm" ]]; then
 	        
+	        shpm_log "   WARN: sh-pm updating itself ..."
 	        if [[ -f $ROOT_DIR_PATH/shpm.sh ]]; then
+	        	shpm_log " - backup actual sh-pm version to $ROOT_DIR_PATH/tmpoldshpm ..."
 	        	if [[ ! -d $ROOT_DIR_PATH/tmpoldshpm ]]; then
 		        	mkdir $ROOT_DIR_PATH/tmpoldshpm		
 				fi
 	        	mv $ROOT_DIR_PATH/shpm.sh $ROOT_DIR_PATH/tmpoldshpm
 	        fi
 	        
-	        TARGET_FOLDER=$ARTIFACT_ID"-"$VERSION
-	        cp $LIB_DIR_PATH/$TARGET_FOLDER/shpm.sh	$ROOT_DIR_PATH
+        	TARGET_FOLDER=$ARTIFACT_ID"-"$VERSION
+        	
+	        if [[ -f $LIB_DIR_PATH/$TARGET_FOLDER/shpm.sh ]]; then
+	        	shpm_log " - update shpm.sh ..."
+	        	cp $LIB_DIR_PATH/$TARGET_FOLDER/shpm.sh	$ROOT_DIR_PATH
+	        fi
 		fi
-
+		
+		shpm_log "Update $DEP_ARTIFACT_ID to $DEP_VERSION: Finish"
 }
 
 build_release() {
+
+	run_all_tests
 
 	local HOST=${REPOSITORY[host]}
 	local PORT=${REPOSITORY[port]}	
@@ -188,7 +279,7 @@ build_release() {
 	cd $TARGET_DIR_PATH
 	cp $ROOT_DIR_PATH/bootstrap.sh $TARGET_DIR_PATH/$TARGET_FOLDER
 	
-	tar -cvzf $TARGET_FOLDER".tar.gz" $TARGET_FOLDER
+	tar -czf $TARGET_FOLDER".tar.gz" $TARGET_FOLDER
 	
 	if [[ -d $TARGET_DIR_PATH/$TARGET_FOLDER ]]; then
 		rm -rf $TARGET_DIR_PATH/$TARGET_FOLDER	
@@ -200,24 +291,32 @@ build_release() {
 
 publish_release() {
 
+	clean_release
+	build_release
+
 	local HOST=${REPOSITORY[host]}
 	local PORT=${REPOSITORY[port]}	
 
-	TARGET_FOLDER=$ARTIFACT_ID"-"$VERSION
-	FILE_PATH=$TARGET_DIR_PATH/$TARGET_FOLDER".tar.gz" 
+	local TARGET_FOLDER=$ARTIFACT_ID"-"$VERSION
+	local FILE_PATH=$TARGET_DIR_PATH/$TARGET_FOLDER".tar.gz" 
 	
 	curl -F file=@"$FILE_PATH" http://$HOST:$PORT/sh-archiva/snapshot/$GROUP_ID/$ARTIFACT_ID/$VERSION
 }
 
 run_all_tests() {
+
+	local ACTUAL_DIR=$(pwd)
+
 	cd $TEST_DIR_PATH
 	
-	for file in `ls *_test.sh`
+	local TEST_FILES=( $(ls *_test.sh 2> /dev/null) );
+	
+	for file in ${TEST_FILES[@]}
 	do
 		source $file
 	done
 	
-	cd $ROOT_DIR_PATH
+	cd $ACTUAL_DIR
 }
 
 auto_update() {
@@ -226,27 +325,14 @@ auto_update() {
 	local PORT=${REPOSITORY[port]}	
 	
 	for DEP_ARTIFACT_ID in "${!DEPENDENCIES[@]}"; do 
-	
 	    if [[ "$DEP_ARTIFACT_ID" == "sh-pm" ]]; then		
-			
-			echo "Updating sh-pm ..."
-			
 			update_dependency $DEP_ARTIFACT_ID
-	        
-	        if [[ ! -d $ROOT_DIR_PATH/tmpoldshpm ]]; then
-		        mkdir $ROOT_DIR_PATH/tmpoldshpm		
-			fi
-	        mv $ROOT_DIR_PATH/shpm.sh $ROOT_DIR_PATH/tmpoldshpm
-	        
-	        TARGET_FOLDER=$ARTIFACT_ID"-"$VERSION
-	        cp $LIB_DIR_PATH/$TARGET_FOLDER/shpm.sh	$ROOT_DIR_PATH
-	        
 	        exit 0    
 	    fi
 	done
 	
 	echo "Could not update sh-pm: sh-pm not present in dependencies of pom.sh"
-	exit 1
+	exit 1004
 }
